fix: `int` -> `long` in claim block events

+++ common/src/main/java/net/william278/huskclaims/event/ClaimBlocksChangeEvent.java
@@ -28,14 +28,14 @@ public interface ClaimBlocksChangeEvent extends CancellableEvent {
     @NotNull
     User getUser();
 
-    int getOldBlocks();
+    long getOldBlocks();
 
-    int getNewBlocks();
+    long getNewBlocks();
 
     @NotNull
     UserManager.ClaimBlockSource getReason();
 
-    default int getClaimBlocksChange() {
+    default long getClaimBlocksChange() {
         return getNewBlocks() - getOldBlocks();
     }
 }
